% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovMats-data.R
\docType{data}
\name{CovMats}
\alias{CovMats}
\title{Data to run with run_EPR_NERR_simple.R.  CovMats has 7 covariance matrices for the seven NERR sites.
CovMats is the second input into run_EPR_NERR_simple}
\format{Two objects of class list}
\usage{
data(CovMats)
}
\description{
Data to run with run_EPR_NERR_simple.R.  CovMats has 7 covariance matrices for the seven NERR sites.
CovMats is the second input into run_EPR_NERR_simple
}
\examples{
load("CovMats.Rdata")

}
\keyword{datasets}
