% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernmatsimple.R
\name{kernmatsimple}
\alias{kernmatsimple}
\title{Internal middle man function to help create kernel}
\usage{
kernmatsimple(x, F, M, fixparm, T)
}
\arguments{
\item{x}{This is Params$x from EPR_NERR_simple}

\item{F}{This is the fishing rate, set to 0 in EPR_NERR_simple}

\item{M}{This is the background non-fishing mortality rate}

\item{fixparm}{This is Params that has Params$Linf, Params$k (only one value for each) but there are other components added in EPR_NERR_simple before Params goes into this function}

\item{T}{This is the "step size", set to 1 in EPR_NERR_simple}
}
\description{
This is an internal middle man function that creates x and y matrices that feed into mkkernsimple.  Not meant to be run on its own.
}
\examples{
kernmatsimple(x,F,fixparm,T)
}
\keyword{NERR}
\keyword{function}
\keyword{helper}
\keyword{kernel}
