% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkkernsimple.R
\name{mkkernsimple}
\alias{mkkernsimple}
\title{Internal function to create kernel}
\usage{
mkkernsimple(x, y, F, M, fixparm, T)
}
\arguments{
\item{x}{This is a matrix x that comes from kernmatsimple}

\item{y}{This is a matrix y that comes from kernmatsimple}

\item{F}{This is the fishing rate, set to 0 in EPR_NERR_simple}

\item{M}{This is the background non-fishing mortality rate}

\item{fixparm}{This is Params that has Params$M, Params$Linf, Params$k (only one value for each) but there are other components added in EPR_NERR_simple before Params goes into this function}

\item{T}{This is the "step size", set to 1 in EPR_NERR_simple}
}
\description{
This is an internal function that creates the growth kernel that gets used in EPR_NERR_simple.  Not meant to be run on its own.
}
\examples{
mkkernsimple(x,y,F,fixparm,T)
}
\keyword{NERR}
\keyword{kernel}
