% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamsInfo.R
\docType{data}
\name{Params}
\alias{Params}
\title{Data to run with run_EPR_NERR_simple.R.  Params is a list with 4 components, 
Sitenames, M, Linf, and k.  Each one of these variables has 
7 entries for the 7 NERR sites.  Params is one of the inputs to run_EPR_NERR_simple.}
\format{Two objects of class list}
\usage{
data(Params)
}
\description{
Data to run with run_EPR_NERR_simple.R.  Params is a list with 4 components, 
Sitenames, M, Linf, and k.  Each one of these variables has 
7 entries for the 7 NERR sites.  Params is one of the inputs to run_EPR_NERR_simple.
}
\examples{
load("Params.Rdata")

}
\keyword{datasets}
